requirements.txt
fastapi
uvicorn
httpx
openai>=1.0.0
pillow
python-multipart